test(moe): move mmap-release test into test_llm.py so CI collects it
